refactor: Move scssTypesPlugin to a separate file
