+++ esbuild.config.mjs
@@ -1,11 +1,10 @@
 import builtins from "builtin-modules";
-import chokidar from "chokidar";
 import esbuild from "esbuild";
 import { postcssModules, sassPlugin } from "esbuild-sass-plugin";
 import fs from "fs";
 import process from "process";
-import Tsm from "typed-scss-modules";
 import { lezerPlugin } from "./esbuild-lezer-plugin.mjs";
+import { scssTypesPlugin } from "./esbuild-scss-types-plugin.mjs";
 
 const banner = `/*
 THIS IS A GENERATED/BUNDLED FILE BY ESBUILD
@@ -14,41 +13,6 @@ if you want to view the source, please visit the github repository of this plugi
 `;
 
 const prod = process.argv[2] === "production";
-
-// generates and regenerates scss type declarations
-export const scssTypesPlugin = (stylesPattern, watch) => {
-    const generateScssTypeDeclarations = async (pattern, options) => {
-        console.log("GENERATE SCSS TYPES", pattern, options, Tsm);
-        await Tsm.default(pattern, options);
-    };
-
-    let setupDone = false;
-
-    return {
-        name: "scss-types-plugin",
-        setup(build) {
-            console.log("setup", stylesPattern);
-            build.onStart(async () => {
-                console.log("onstart");
-                await generateScssTypeDeclarations(stylesPattern, {  });
-                setupDone = true;
-            });
-
-            if (watch) {
-                console.log("setup watch");
-                chokidar.watch(stylesPattern).on(
-                    "all", // was on(change)
-                    async (event, filename) => {
-                        console.log("onwatch", event, filename);
-                        if (event == "add" && !setupDone) return;
-                        if (!(event == "change" || event == "add")) return;
-                        await generateScssTypeDeclarations(filename);
-                    }
-                );
-            }
-        },
-    };
-};
 
 // renames `main.css(.map)` to `styles.css(.map)` in build dir
 const renameCompiledCSS = {
